[SDK:REACTOS] Move our ROS-specific header-only UI libraries into a "ui" sub-folder. (#4205)
