+++ dll/shellext/acppage/precomp.h
@@ -17,7 +17,7 @@
 #include <atlsimpcoll.h>
 #include <atlstr.h>
 #include <atlwin.h>
-#include <rosdlgs.h>
+#include <ui/rosdlgs.h>
 
 ULONG DbgPrint(PCH Format,...);
 #include <apphelp.h>
