+++ dll/win32/shell32/shellmenu/shellmenu.h
@@ -52,7 +52,7 @@
 #include <undocuser.h>
 
 #include <shellutils.h>
-#include <rosctrls.h>
+#include <ui/rosctrls.h>
 #include "../shresdef.h"
 
 #include <wine/debug.h>
