+++ sdk/include/reactos/ui/layout.h
