+++ sdk/include/reactos/ui/rosdlgs.h
